Refactor RichTextEditor to use react-editor-js wrapper
