+++ src/components/RichTextEditor/consts.ts
@@ -1,0 +1,47 @@
+import EditorJS, {
+  LogLevels,
+  OutputData,
+  ToolConstructable,
+  ToolSettings
+} from "@editorjs/editorjs";
+import Embed from "@editorjs/embed";
+import Header from "@editorjs/header";
+import List from "@editorjs/list";
+import Paragraph from "@editorjs/paragraph";
+import Quote from "@editorjs/quote";
+import strikethroughIcon from "@saleor/icons/StrikethroughIcon";
+import classNames from "classnames";
+import createGenericInlineTool from "editorjs-inline-tool";
+const inlineToolbar = ["link", "bold", "italic", "strikethrough"];
+
+export const tools: Record<string, ToolConstructable | ToolSettings> = {
+  embed: Embed,
+  header: {
+    class: Header,
+    config: {
+      defaultLevel: 1,
+      levels: [1, 2, 3]
+    },
+    inlineToolbar
+  },
+  list: {
+    class: List,
+    inlineToolbar
+  },
+  quote: {
+    class: Quote,
+    inlineToolbar
+  },
+  paragraph: {
+    class: Paragraph,
+    inlineToolbar
+  },
+  strikethrough: createGenericInlineTool({
+    sanitize: {
+      s: {}
+    },
+    shortcut: "CMD+S",
+    tagName: "s",
+    toolboxIcon: strikethroughIcon
+  })
+};
